requirements.txt
setuptools==80.9.0
pywebview==6.1
pyinstaller==6.16.0; sys_platform != 'darwin'
py2app==0.28.8; sys_platform == 'darwin'
ruff==0.14.3
pyflow-ts==0.2.4
numpy==2.3.4